Moving getConversionRate code to util file
